conn/display: simplify Drawer
display: move devices.Display as display.Drawer

This removes the last interface from devices, which was misplaced due to
historical accident. The new package display will also be the location for an
interface for text output only devices.

Remove io.Writer from this interface. While it's a good performance optimization
for some drivers, it shouldn't be required.

Change Draw():
- Remove destination Rectangle from Draw(), not covering the whole display is
  now undefined, it is not a requirement for the driver to buffer the whole
  display.
- Return an error, so that communication erorr can be surfaced correctly,
  instead of an adhoc driver specific Err() method.